add detail anime screen

+++ app/src/main/java/id/isato/emina/ui/navigation/Screen.kt
@@ -4,6 +4,6 @@ sealed class Screen(val route: String) {
     object Home : Screen("home")
     object Profile : Screen("profile")
     object DetailAnime : Screen("home/{animeId}") {
-        fun createRoute(animeId: Long) = "home/$animeId"
+        fun createRoute(animeId: Int) = "home/$animeId"
     }
 }
